+++ datumaro/components/extractor.py
@@ -5,7 +5,7 @@
 
 from enum import Enum
 from glob import iglob
-from typing import Iterable, List, Dict, Optional
+from typing import Callable, Iterable, List, Dict, Optional
 import numpy as np
 import os
 import os.path as osp
@@ -668,13 +668,39 @@ class Importer:
         return sources
 
     @classmethod
-    def _find_sources_recursive(cls, path, ext, extractor_name,
-            filename='*', dirname='', file_filter=None, max_depth=3):
+    def _find_sources_recursive(cls, path: str, ext: Optional[str],
+            extractor_name: str, filename: str = '*', dirname: str = '',
+            file_filter: Optional[Callable[[str], bool]] = None,
+            max_depth: int = 3):
+        """
+        Finds sources in the specified location, using the matching pattern
+        to filter file names and directories.
+        Supposed to be used, and to be the only call in subclasses.
 
-        if (path.endswith(ext) and osp.isfile(path)) or \
-                (not ext and osp.isdir(path) and dirname and \
-                os.sep + osp.normpath(dirname) + os.sep in \
-                    osp.abspath(path) + os.sep):
+        Paramters:
+        - path - a directory or file path, where sources need to be found.
+        - ext - file extension to match. To match directories,
+            set this parameter to None or ''. Comparison is case-independent,
+            a starting dot is not required.
+        - extractor_name - the name of the associated Extractor type
+        - filename - a glob pattern for file names
+        - dirname - a glob pattern for filename prefixes
+        - file_filter - a callable (abspath: str) -> bool, to filter paths found
+        - max_depth - the maximum depth for recursive search.
+
+        Returns: a list of source configurations
+            (i.e. Extractor type names and c-tor parameters)
+        """
+
+        if ext:
+            if not ext.startswith('.'):
+                ext = '.' + ext
+            ext = ext.lower()
+
+        if (ext and path.lower().endswith(ext) and osp.isfile(path)) or \
+                (not ext and dirname and osp.isdir(path) and \
+                os.sep + osp.normpath(dirname.lower()) + os.sep in \
+                    osp.abspath(path.lower()) + os.sep):
             sources = [{'url': path, 'format': extractor_name}]
         else:
             sources = []